Fixed a precision issue that was present.
We need to be able to reach 100 choose 50 which requires a HUGE amount of
precision. It appears that long double, 2^80 precision, is the most C will
allow and hopefully that is enough...

+++ test/combinations_test.c
@@ -5,10 +5,10 @@
 int
 main ( void )
 {
-  unsigned long long c ;
+  long double c ;
   c = n_choose_k ( 12 , 9 ) ;
-  printf ( "%i\n" , c ) ;
-  if ( c == 220 )
+  printf ( "%.0Lf\n" , c ) ;
+  if ( ( int ) c == 220 )
     printf ( "12c9 calculated correctly.\n" ) ;
   else
   {
@@ -16,8 +16,8 @@ main ( void )
     exit ( -1 ) ;
   }
   c = n_choose_k ( 18 , 3 ) ;
-  printf ( "%i\n" , c ) ;
-  if ( c == 816 )
+  printf ( "%.0Lf\n" , c ) ;
+  if ( ( int ) c == 816 )
     printf ( "18c3 calculated correctly.\n" ) ;
   else
   {
@@ -25,8 +25,8 @@ main ( void )
     exit ( -1 ) ;
   }
   c = n_choose_k ( 35 , 8 ) ;
-  printf ( "%i\n" , c ) ;
-  if ( c == 23535820 )
+  printf ( "%.0Lf\n" , c ) ;
+  if ( ( int ) c == 23535820 )
     printf ( "35c8 calculated correctly.\n" ) ;
   else
   {
